update: move dispatch and selector Redux logic to layout page
